Add files via upload
This is the .ipynb file which consists of code for training U-Net. Please note that through this code we trained U-Net for multiple datasets. It also consists of some code that can be used for testing the results of the trained model. This code has been commented out and can be un-commented based on which result is to be visualized.